Detect if Bluetooth adapter is on
This is the first stage of integrating Bluetooth into this app.

If the Bluetooth adapter is off, it does not let the user access
the main app functionality, and instead prompts the user to turn it
on.

In Android, Bluetooth can be turned on in app, so a button which does
this is presented, however for iOS, this is not possible, so the user
must turn it on in Settings.

I followed the "Bluetooth On and Off" section of the tutorial found
here: https://pub.dev/packages/flutter_blue_plus