Handle Bluetooth connection lifecycle
For simplicity, I made the decision that you cannot have an
existing connection while scanning for another (ie, when you
are connected to a device, beginning to scan for another means
that the existing connection will be destroyed).

After all, this project is really only a proof of concept/experiment

This commit allows the user to actually connect to the Bluetooth
devices which come up on the scan.

Also decided to conditionally show the UI elements which will
send/receive data to the connected device, only when there is
a connected device

I also set up a listener for Bluetooth disconnect events to
hide these elements once the device becomes disconnected