Implement sending and receiving of data over BLE
This is done by finding a service which the device offers
which has a writeWithoutResponse and notify characteristic,
then using this to write to and read from the peripheral.

In the case of the HM-10 I am experimenting with, the only
service it offers which has such a characteristic is the serial
/UART service which is the one I am interested in.

There may be other devices which advertise many services which have
such characteristics, in which case the first one encountered is
chosen for simplicity's sake (again, this being a proof of
concept/experiment which is meant to inform later projects).

In the case where the device offers no services which have such
characteristics, the connection will fail.

In order to test this properly, I will need to implement the server
on the HM-10 peripheral, which will do something like reverse the
received string and return it (I have not decided yet). This will be
in a different repository.